tests/wpt/tests/html/canvas/element/path-objects/2d.path.arc.scale.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.arc.scale.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.arc.scale.1</h1>
<p class="desc">Non-uniformly scaled arcs are the right shape</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Non-uniformly scaled arcs are the right shape");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.scale(2, 0.5);
  ctx.fillStyle = '#0f0';
  ctx.beginPath();
  ctx.arc(25, 50, 56, 0, 2*Math.PI, false);
  ctx.fill();
  ctx.fillStyle = '#f00';
  ctx.beginPath();
  ctx.moveTo(-25, 50);
  ctx.arc(-25, 50, 24, 0, 2*Math.PI, false);
  ctx.moveTo(75, 50);
  ctx.arc(75, 50, 24, 0, 2*Math.PI, false);
  ctx.moveTo(25, -25);
  ctx.arc(25, -25, 24, 0, 2*Math.PI, false);
  ctx.moveTo(25, 125);
  ctx.arc(25, 125, 24, 0, 2*Math.PI, false);
  ctx.fill();

  _assertPixel(canvas, 0,0, 0,255,0,255);
  _assertPixel(canvas, 50,0, 0,255,0,255);
  _assertPixel(canvas, 99,0, 0,255,0,255);
  _assertPixel(canvas, 0,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 99,25, 0,255,0,255);
  _assertPixel(canvas, 0,49, 0,255,0,255);
  _assertPixel(canvas, 50,49, 0,255,0,255);
  _assertPixel(canvas, 99,49, 0,255,0,255);

});
</script>
